refactor: Изменил структуру папок
