+++ src/components/ui/LogoLink/LogoLink.jsx
@@ -1,6 +1,6 @@
 import { NavLink } from 'react-router-dom';
 
-import Logo from '../Logo/Logo';
+import Logo from '../ui/Logo/Logo';
 
 import './LogoLink.css';
