+++ src/pages/Landing/Landing.jsx
@@ -3,7 +3,7 @@ import AboutProject from './AboutProject/AboutProject';
 import Techs from './Techs/Techs';
 import AboutMe from './AboutMe/AboutMe';
 import Portfolio from './Portfolio/Portfolio';
-import Preloader from '../../components/Preloader/Preloader';
+import Preloader from '../../components/ui/Preloader/Preloader';
 
 const Landing = () => {
   return (
